+++ src/Product/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'Vanilo Product Module',
-    'version' => '3.2.0'
+    'version' => '3.3.0'
 ];