+++ src/MasterProduct/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2022-XX-YY
 
 - Added the `MasterProduct::actives()` and `inactives()` query builder scopes
+- Changed minimum Concord version requirement to v1.12
 
 ## 3.1.0
 ##### 2022-11-07